requirements.txt
pandas==2.3.3          
openpyxl==3.1.2        
PyYAML==6.0           
requests==2.31.0       